Extract manage staging site card content to separate component
